# docs/swagger.yaml
definitions:
...
  api.assignVehicleRequest:
    properties:
      user_id:
        type: integer
      vehicle_id:
        type: integer
    type: object
...
  api.createVehicleRequest:
    properties:
      assigned_driver:
        type: integer
      car_model:
        type: string
      color:
        type: string
      current_mileage:
        type: integer
      last_maintenance:
        type: string
      license_plate:
        type: string
      make:
        type: string
      next_maintenance:
        type: string
      notes:
        type: string
      sitting_capacity:
        type: integer
      status:
        type: string
      type:
        type: string
      vin:
        type: string
      year:
        type: integer
    type: object
  api.createVehicleResponse:
    properties:
      ID:
        type: integer
      assigned_driver:
        type: integer
      car_model:
        type: string
      color:
        type: string
      current_mileage:
        type: integer
      last_maintenance:
        type: string
      license_plate:
        type: string
      make:
        type: string
      next_maintenance:
        type: string
      notes:
        type: string
      sitting_capacity:
        type: integer
      status:
        type: string
      type:
        type: string
      vin:
        type: string
      year:
        type: integer
    type: object
...
  api.registerVehicleRequest:
    properties:
      assigned_driver:
        type: integer
      car_model:
        type: string
      color:
        type: string
      current_mileage:
        type: integer
      last_maintenance:
        type: string
      license_plate:
        type: string
      make:
        type: string
      next_maintenance:
        type: string
      notes:
        type: string
      sitting_capacity:
        type: integer
      type:
        type: string
      vin:
        type: string
      year:
        type: integer
    type: object
  api.userResponse:
    properties:
      ID:
        type: integer
      address:
        type: string
      driving_license_number:
        type: string
      email:
        type: string
      first_name:
        type: string
      goverment_id:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      phone_number:
        type: string
      role:
        type: string
      status:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This API serves as a backend for Vehicle Management System
  title: Vehicle Management System API
  version: "1.0"
paths:
...
  /vehicle:
    get:
      description: Get all vehicles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/api.createVehicleResponse'
              type: array
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get all vehicles
      tags:
      - vehicle
    post:
      consumes:
      - application/json
      description: Create a vehicle
      parameters:
      - description: Vehicle
        in: body
        name: vehicle
        required: true
        schema:
          $ref: '#/definitions/api.createVehicleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.createVehicleResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a vehicle
      tags:
      - vehicle
  /vehicle/{id}:
    delete:
      description: Delete a vehicle
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Delete a vehicle
      tags:
      - vehicle
    get:
      description: Get a vehicle
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.createVehicleResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a vehicle
      tags:
      - vehicle
    post:
      description: Changes status of pending vehicles to Active
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.createVehicleResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Activate pending vehicle
      tags:
      - vehicle
    put:
      consumes:
      - application/json
      description: Update a vehicle
      parameters:
      - description: Vehicle ID
        in: path
        name: id
        required: true
        type: integer
      - description: Vehicle
        in: body
        name: vehicle
        required: true
        schema:
          $ref: '#/definitions/api.createVehicleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.createVehicleResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a vehicle
      tags:
      - vehicle
  /vehicle/assign:
    post:
      consumes:
      - application/json
      description: Assign a vehicle to a driver
      parameters:
      - description: Vehicle
        in: body
        name: vehicle
        required: true
        schema:
          $ref: '#/definitions/api.assignVehicleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.userResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Assign a vehicle to a driver
      tags:
      - vehicle
  /vehicle/register:
    post:
      consumes:
      - application/json
      description: Register a vehicle
      parameters:
      - description: Vehicle
        in: body
        name: vehicle
        required: true
        schema:
          $ref: '#/definitions/api.registerVehicleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.createVehicleResponse'
      security:
      - ApiKeyAuth: []
      summary: Register a vehicle
      tags:
      - vehicle
  /vehicle/unassign:
    post:
      consumes:
      - application/json
      description: Unassign a vehicle from a driver
      parameters:
      - description: Vehicle
        in: body
        name: vehicle
        required: true
        schema:
          $ref: '#/definitions/api.assignVehicleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.userResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Unassign a vehicle from a driver
      tags:
      - vehicle
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"